CRITICAL: Fix FAISS dtype error - convert embeddings to float32 before normalization
Error: TypeError in fvec_renorm_L2, argument 3 of type 'float *'
Root cause: FastEmbed embeddings weren't explicitly float32
Fix: Force dtype=np.float32 in np.array() BEFORE faiss.normalize_L2()

This was causing index build to fail after 15 minutes of embedding generation.

+++ ine_mcp_python/ine_mcp/embeddings/search.py
@@ -97,18 +97,19 @@ class SemanticSearchEngine:
         # Generate embeddings using FastEmbed
         logger.info("Generating embeddings...")
         # FastEmbed returns a generator, convert to numpy array
+        # CRITICAL: Must be float32 for FAISS compatibility
         embeddings_list = list(self.model.embed(texts, batch_size=32))
-        embeddings = np.array(embeddings_list)
+        embeddings = np.array(embeddings_list, dtype=np.float32)
 
         # Create FAISS index
         dimension = embeddings.shape[1]
         index = faiss.IndexFlatIP(dimension)  # Inner Product for cosine similarity
 
-        # Normalize embeddings for cosine similarity
+        # Normalize embeddings for cosine similarity (requires float32)
         faiss.normalize_L2(embeddings)
 
-        # Add to index
-        index.add(embeddings.astype(np.float32))
+        # Add to index (already float32, no conversion needed)
+        index.add(embeddings)
 
         self._index = index
         self._metadata = metadata
